Add support to configure custom routes to TGWs
Adding a new variable custom_route_to_tgw which is map of objects.
The map key is used to identify required subnets.
Each object contains route configuration.
Right now I'm creating a new route in the private and public subnets.
Not sure if this can be useful also for the transit_gateway and / or core_network subnets.
It is fair to mention that there is one limitation of this variable - one cannot create more routes per a subnet.
Not sure how to achieve it without too much changes.