Configure additional routes for the subnets
Adding new optional attribute routes into variable subnets.
This is useful in environments where core network, transit
gateways, etc. are created out of operator control.

Notes to the implementation:
We are limited here by Terraform because for_each works only with
set or map. As we have multiple subnets for a subnet / route table
we cannot use it. For that reason the code creates list of maps where
each map represents a route for the specific route table (attribute
route_table_name). In such case we can use count to iterate over it.

We are limited here by Terraform because for_each works only with
a set or map. As we have multiple routes for a subnet / route table,
we cannot use it. For that reason, the code creates a list of maps
where each map represents a route for the specific route table
(attribute route_table_name). In such case, we can use count to
iterate over it.